create file
For use on phoenix. Takes in a file of known L1 regions and the corresponding genome. Outputs a random sample of 100 L1s (between 4.5kb-5.5kb in length) and 100 non-L1 genomic pieces (similar lengths), in FASTA format. Does not allow any of the output sequences to have unknown nucleotides (i.e. Ns).